\docType{methods}
\name{dbDataType,H2Object-method}
\alias{dbDataType,H2Object-method}
\title{Determine the SQL Data Type of an R object.}
\usage{
\S4method{dbDataType}{H2Object}(dbObj, obj, ...)
}
\arguments{
  \item{dbObj}{a \code{\linkS4class{H2Object}} object.}

  \item{obj}{an R object whose SQL type we want to
  determine.}

  \item{...}{Needed for compatibility with generic.
  Otherwise ignored.}
}
\description{
Determine the SQL Data Type of an R object.
}

