% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/H2Connection.R
\docType{methods}
\name{dbGetInfo,H2Connection-method}
\alias{dbGetInfo,H2Connection-method}
\title{Get metadata about a H2Connection object.}
\usage{
\S4method{dbGetInfo}{H2Connection}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object of class \code{\linkS4class{H2Connection}}}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Get metadata about a H2Connection object.
}

