\docType{class}
\name{H2Result-class}
\alias{H2Result-class}
\alias{dbHasCompleted,H2Result-method}
\title{Class H2Result}
\usage{
\S4method{dbHasCompleted}{H2Result}(res, ...)
}
\arguments{
  \item{res}{ refers to a \code{DBIResult} object. }

  \item{...}{ any driver-specific arguments.  }
}
\description{
Class H2Result
}

