% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2Driver.R
\docType{methods}
\name{dbConnect,H2Driver-method}
\alias{dbConnect,H2Driver-method}
\title{Connect to a h2 database.}
\usage{
\S4method{dbConnect}{H2Driver}(drv, url = "mem:", user = "sa",
  password = "", ...)
}
\arguments{
\item{drv}{the database driver (\code{\link{H2}})}

\item{url}{the url to connect to (Visit \url{http://h2database.com/html/features.html#database_url} for a reference).
If the url is a path to a local file the '.h2.db' suffix is stripped off automatically if present.}

\item{user}{the user to log in}

\item{password}{the users password}

\item{...}{further arguments passed to \code{\linkS4class{JDBCConnection}}.}
}
\description{
Connect to a h2 database.
}

