% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2Driver.R
\docType{methods}
\name{dbConnect,H2Driver-method}
\alias{dbConnect,H2Driver-method}
\title{Connect to an H2 database.}
\usage{
\S4method{dbConnect}{H2Driver}(drv, url = "mem:", user = "sa",
  password = "", sql_dialect = h2_dialect, ...)
}
\arguments{
\item{drv}{a \code{\linkS4class{JDBCConnection}} object}

\item{url}{the url to connect to (Visit \url{http://h2database.com/html/features.html#database_url} for a reference).
If the url is a path to a local file the '.h2.db' suffix is stripped off automatically if present.}

\item{password}{the password for the connection}

\item{...}{Ignored. Included for compatibility with generic.}
}
\description{
Connect to an H2 database.
}

