% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/H2Driver.R
\docType{class}
\name{H2Driver-class}
\alias{H2}
\alias{H2Driver-class}
\title{Class H2Driver and factory method H2.}
\usage{
H2(...)
}
\arguments{
\item{...}{arguments to \code{\link[RJDBC]{JDBC}}}
}
\description{
Class H2Driver and factory method H2.
}

