% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{H2Driver-class}
\alias{H2}
\alias{H2Driver-class}
\title{Class H2Driver and factory method H2.}
\usage{
H2(...)
}
\arguments{
\item{...}{arguments to \link[RJDBC]{JDBC}}
}
\description{
Class H2Driver and factory method H2.
}

