% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2Driver.R
\docType{class}
\name{H2Driver-class}
\alias{H2}
\alias{H2Driver-class}
\alias{driver}
\title{Class H2Driver and factory method H2.}
\usage{
driver(...)

H2(...)
}
\arguments{
\item{...}{arguments to \code{\link[dbj]{driver}}}

\item{...}{arguments to \code{\link[dbj]{driver}}}
}
\description{
Class H2Driver and factory method H2.

Deprecated. Use dbj.h2::driver()
}
\keyword{internal}

